Fixed #8: Replaced generating an ld.conf file by changing the RPATH of binaries
The RPATH generated by the compiler point (absolutely) to the rpmuiild BUILD
directory. To workaround this one has added an ld.conf file which contains the
path to the custom Qt library.

While this works fine for headless systems where no Qt is required by default,
this causes problems on headed systems (in fact it caused problems with KDE on
Fedora 17).

This commit changes the the RPATH to point to the library directory where Qt
is located.

Building this package would require QA_RPATHS=$[ 0x0004 ], use it this way:

   QA_RPATHS=$[ 0x0004 ] rpmbuild -bXYZ ...